requirements.txt
django-leaflet
django-geojson
Pillow
jsonfield